app/requirements.txt
## CORE (REQUIRED to be stable packages) ##
gunicorn==19.4.5                                # python 3
Django==1.8.9                                   # python 3
djangorestframework==3.4.0                      # python 3
django-tastypie==0.12.2                         # python 3
PyMySQL==0.6.7                                  # python 3
pyrax==1.9.7                                    # python 2 x
Pillow==2.6.2                                   # python 3
celery==3.1.19                                  # python 3
redis==2.10.5                                   # python 3
python-etcd==0.4.3                              # python 3

## SECONDARY (SHOULD be stable packages) ##
markdown
envitro>=0.4.2,<1.0.0                           # python 3
django-grappelli>=2.7.3,<3.0.0                  # python 3
jsonfield>=1.0.3                                # python 3
django-uuidfield>=0.5.0                         # python 3
django-bitfield>=1.8.0                          # python 3
pytz>=2015.7                                    # python 3
python-dateutil>=2.4.2                          # python 3
MonthDelta>=1.0b,<2.0                           # python 3
raven>=5.10.2,<6.0                              # python 3
unicodecsv>=0.14.1,<1.0.0                       # python 3

## Third-Party Services ##
stripe>=1.28.0,<1.29.0                          # python 3
sendwithus>=1.7.0,<2.0.0                        # python 3
newrelic>=2.56.0.42                             # python 3

## Testing ##
nose==1.3.7
django-nose==1.4.3
